% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_voices.R
\docType{data}
\name{available_voices}
\alias{available_voices}
\title{List of all the voices for the speak function}
\format{A vector}
\source{
MacOS command line: 'say'
}
\usage{
available_voices
}
\description{
A complete list of all the voices supported by the \code{\link{speak}} function.
}
\keyword{datasets}
