% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsay.R
\docType{package}
\name{rsay-package}
\alias{rsay}
\alias{rsay-package}
\title{Make R on your Mac speak}
\description{
rsay is a simple wrapper around the MacOS command line tool, 'say'.
}
\details{
The only function you're likely to need from \pkg{rsay} is
\code{\link{speak}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sellorm/rsay}
  \item Report bugs at \url{https://github.com/sellorm/rsay/issues}
}

}
\author{
\strong{Maintainer}: Mark Sellors \email{rstuff@5vcc.com}

}
