% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speak.R
\name{speak}
\alias{speak}
\title{Speak your input text}
\usage{
speak(input, voice)
}
\arguments{
\item{input}{A string}

\item{voice}{An optional string matching a voice

The complete list of voices can be found in available_voices}
}
\description{
Speak your input text
}
\examples{
\dontrun{
speak("hello world")
speak("Hello world", voice = "Samantha")
}
}
